ADK_Agent/requirements.txt
elevenlabs
google-adk>=1.3.0,<2.0.0
dotenv >=0.9.9,<0.10.0
pendulum >=3.1.0,<4.0.0